32 bit blocks

+++ text-based-burger/font_loader.h
@@ -10,4 +10,4 @@
 // Loads the font file and returns a list of vec2s containing the upper and lower halves of the characters as a 
 // 64 bit number representation. If your dont isnt 8*16 black and white you will need to modify this function.
 
-std::vector<uint64_t> load_font(const std::string& font_path);
+std::vector<uint32_t> load_font(const std::string& font_path);
